move replay file to models/files/
